Read slides from `stdin`
Fixes #15

This change allows users to pipe in slides from `stdin` which allows for
`curl`ing remote slides and presenting them as well as piping multiple files
into one (with `cat`) and much more.

```
curl http://example.com/slides.md | slides
```

This change also maintains backwards compatibility with the current API which
allows for users to simply specify a file name. In the future, it would be nice
to remove that in favour of simply reading the file from stdin.

Like so:
```
slides < examples/slides.md
```