code: Always display the output of code blocks
Any output that is printed may still be useful even if the command returns
an error so lets display it. Printing out even in case of error is fine,
if there's no output and output.Write handles that fine. I tested this by
trying to run a code block for a language I do not have the compiler for.